refactoring Python API
