--- docs/api/nunavut.cli.rst
@@ -1,18 +1,0 @@
-nunavut.cli
-===========
-
-.. automodule:: nunavut.cli
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Submodules
-----------
-
-nunavut.cli.runners
-~~~~~~~~~~~~~~~~~~~
-
-.. automodule:: nunavut.cli.runners
-   :members:
-   :undoc-members:
-   :show-inheritance:
