--- docs/api/nunavut.jinja.rst
@@ -1,34 +1,0 @@
-nunavut.jinja
-=============
-
-.. automodule:: nunavut.jinja
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Submodules
-----------
-
-nunavut.jinja.environment
-~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. automodule:: nunavut.jinja.environment
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-nunavut.jinja.extensions
-~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. automodule:: nunavut.jinja.extensions
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-nunavut.jinja.loaders
-~~~~~~~~~~~~~~~~~~~~~
-
-.. automodule:: nunavut.jinja.loaders
-   :members:
-   :undoc-members:
-   :show-inheritance:
