--- docs/api/nunavut.lang.c.rst
@@ -1,16 +1,0 @@
-nunavut.lang.c
-==============
-
-.. automodule:: nunavut.lang.c
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   nunavut.lang.c.support
-   nunavut.lang.c.templates
