--- docs/api/nunavut.lang.c.support.rst
@@ -1,7 +1,0 @@
-nunavut.lang.c.support
-======================
-
-.. automodule:: nunavut.lang.c.support
-   :members:
-   :undoc-members:
-   :show-inheritance:
