--- docs/api/nunavut.lang.c.templates.rst
@@ -1,7 +1,0 @@
-nunavut.lang.c.templates
-========================
-
-.. automodule:: nunavut.lang.c.templates
-   :members:
-   :undoc-members:
-   :show-inheritance:
