--- docs/api/nunavut.lang.cpp.rst
@@ -1,16 +1,0 @@
-nunavut.lang.cpp
-================
-
-.. automodule:: nunavut.lang.cpp
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   nunavut.lang.cpp.support
-   nunavut.lang.cpp.templates
