--- docs/api/nunavut.lang.cpp.support.rst
@@ -1,7 +1,0 @@
-nunavut.lang.cpp.support
-========================
-
-.. automodule:: nunavut.lang.cpp.support
-   :members:
-   :undoc-members:
-   :show-inheritance:
