--- docs/api/nunavut.lang.cpp.templates.rst
@@ -1,7 +1,0 @@
-nunavut.lang.cpp.templates
-==========================
-
-.. automodule:: nunavut.lang.cpp.templates
-   :members:
-   :undoc-members:
-   :show-inheritance:
