--- docs/api/nunavut.lang.html.rst
@@ -1,15 +1,0 @@
-nunavut.lang.html
-=================
-
-.. automodule:: nunavut.lang.html
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   nunavut.lang.html.templates
