--- docs/api/nunavut.lang.html.templates.rst
@@ -1,7 +1,0 @@
-nunavut.lang.html.templates
-===========================
-
-.. automodule:: nunavut.lang.html.templates
-   :members:
-   :undoc-members:
-   :show-inheritance:
