--- docs/api/nunavut.lang.js.rst
@@ -1,7 +1,0 @@
-nunavut.lang.js
-===============
-
-.. automodule:: nunavut.lang.js
-   :members:
-   :undoc-members:
-   :show-inheritance:
