--- docs/api/nunavut.lang.py.rst
@@ -1,15 +1,0 @@
-nunavut.lang.py
-===============
-
-.. automodule:: nunavut.lang.py
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   nunavut.lang.py.templates
