--- docs/api/nunavut.lang.py.templates.rst
@@ -1,7 +1,0 @@
-nunavut.lang.py.templates
-=========================
-
-.. automodule:: nunavut.lang.py.templates
-   :members:
-   :undoc-members:
-   :show-inheritance:
