--- docs/api/nunavut.lang.rst
@@ -1,19 +1,0 @@
-nunavut.lang
-============
-
-.. automodule:: nunavut.lang
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   nunavut.lang.c
-   nunavut.lang.cpp
-   nunavut.lang.html
-   nunavut.lang.js
-   nunavut.lang.py
