+++ docs/api/nunavut.rst
@@ -3,58 +3,4 @@ nunavut
 
 .. automodule:: nunavut
    :members:
-   :undoc-members:
-   :show-inheritance:
-
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   nunavut.cli
-   nunavut.jinja
-   nunavut.lang
-
-Submodules
-----------
-
-nunavut.dependencies
-~~~~~~~~~~~~~~~~~~~~
-
-.. automodule:: nunavut.dependencies
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-nunavut.generators
-~~~~~~~~~~~~~~~~~~
-
-.. automodule:: nunavut.generators
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-nunavut.postprocessors
-~~~~~~~~~~~~~~~~~~~~~~
-
-.. automodule:: nunavut.postprocessors
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-nunavut.templates
-~~~~~~~~~~~~~~~~~
-
-.. automodule:: nunavut.templates
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-nunavut.version
-~~~~~~~~~~~~~~~
-
-.. automodule:: nunavut.version
-   :members:
-   :undoc-members:
    :show-inheritance:
