+++ setup.cfg
@@ -35,6 +35,7 @@ classifiers =
 package_dir=
     =src
 packages=find:
+package_data={"nunavut": ["py.typed"]}
 install_requires=
     pydsdl ~= 1.16
     pyyaml
