+++ src/nunavut/_version.py
@@ -1,0 +1,15 @@
+#
+# Copyright 2022 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# Copyright (C) 2018-2022  OpenCyphal Development Team  <opencyphal.org>
+# This software is distributed under the terms of the MIT License.
+#
+"""
+
+.. autodata:: __version__
+"""
+
+__version__ = "2.0.0"
+__license__ = "MIT"
+__author__ = "OpenCyphal"
+__copyright__ = "Copyright 2022 Amazon.com, Inc. or its affiliates. All Rights Reserved. Copyright (c) 2022 OpenCyphal."
+__email__ = "maintainers@opencyphal.org"
