--- src/nunavut/version.py
@@ -1,13 +1,0 @@
-#
-# Copyright (C) 2018-2021  UAVCAN Development Team  <uavcan.org>
-# This software is distributed under the terms of the MIT License.
-#
-"""
-
-.. autodata:: __version__
-"""
-
-__version__ = "2.0.0"  #: The version number used in the release of nunavut to pypi.
-
-
-__license__ = "MIT"
