+++ tox.ini
@@ -18,6 +18,7 @@ deps =
 
 [dev]
 deps =
+    {[base]deps}
     autopep8
     rope
     isort
@@ -115,6 +116,9 @@ ignore_missing_imports = True
 [mypy-setuptools.*]
 ignore_missing_imports = True
 
+[mypy-importlib_resources.*]
+ignore_missing_imports = True
+
 # +---------------------------------------------------------------------------+
 # | TOX ENVIRONMENTS
 # +---------------------------------------------------------------------------+
@@ -199,7 +203,7 @@ commands =
 [testenv:lint]
 basepython = python3.10
 deps =
-     {[dev]deps}
+    {[dev]deps}
     black
     flake8
     doc8
@@ -208,7 +212,6 @@ deps =
     lxml
     types-pkg_resources
     types-PyYAML
-    pydsdl
 
 commands =
     flake8 --benchmark --tee --output-file={envtmpdir}/flake8.txt --filename=*.py --exclude=**/jinja2/*,**/markupsafe/* src
